+++ game-digimonstorytimestranger/index.js
@@ -2,8 +2,8 @@
 Name: Digimon Story Time Stranger Vortex Extension
 Structure: Reloaded-II Game (Mod Installer)
 Author: ChemBoy1
-Version: 0.1.1
-Date: 2025-10-09
+Version: 0.1.2
+Date: 2025-10-14
 /////////////////////////////////////////////////*/
 
 //Import libraries
@@ -35,11 +35,15 @@ let DOWNLOAD_FOLDER = '';
 
 //Data for mod types, tools, and installers
 const RELOADED_ID = `${GAME_ID}-reloadedmanager`;
-const RELOADED_NAME = "Reloaded II Mod Manager";
+const RELOADED_NAME = "Reloaded-II Mod Manager";
 const RELOADED_PATH = path.join("Reloaded");
 const RELOADED_EXEC = "reloaded-ii.exe";
+const RELOADED_EXEC_PATH = path.join(RELOADED_PATH, RELOADED_EXEC);
 const RELOADED_URL_LATEST = `https://github.com/Reloaded-Project/Reloaded-II/releases/latest/download/Release.zip`;
 const RELOADED_URL_MANUAL = `https://github.com/Reloaded-Project/Reloaded-II/releases`;
+
+const ELEVATOR_PATH = path.join(util.getVortexPath('application'), 'resources');
+const ELEVATOR_EXEC = "elevate.exe";
 
 const RELOADEDMOD_ID = `${GAME_ID}-reloadedmod`;
 const RELOADEDMOD_NAME = "Reloaded Mod";
@@ -132,10 +136,6 @@ const spec = {
     "names": []
   }
 };
-
-const tools = [
-
-];
 
 // BASIC EXTENSION FUNCTIONS ///////////////////////////////////////////////////
 
@@ -461,11 +461,52 @@ function installReloadedMod(files, fileName) {
   return Promise.resolve({ instructions });
 }
 
+//Installer test for Save (.bin) files
+function testSave(files, gameId) {
+  const isMod = files.some(file => SAVE_EXTS.includes(path.extname(file).toLowerCase()));
+  let supported = (gameId === spec.game.id) && isMod;
+
+  // Test for a mod installer
+  if (supported && files.find(file =>
+      (path.basename(file).toLowerCase() === 'moduleconfig.xml') &&
+      (path.basename(path.dirname(file)).toLowerCase() === 'fomod'))) {
+    supported = false;
+  }
+
+  return Promise.resolve({
+    supported,
+    requiredFiles: [],
+  });
+}
+
+//Installer install Save (.bin) files
+function installSave(files) {
+  const modFile = files.find(file => SAVE_EXTS.includes(path.extname(file).toLowerCase()));
+  const idx = modFile.indexOf(path.basename(modFile));
+  const rootPath = path.dirname(modFile);
+  const setModTypeInstruction = { type: 'setmodtype', value: SAVE_ID };
+
+  // Remove directories and anything that isn't in the rootPath.
+  const filtered = files.filter(file =>
+    ((file.indexOf(rootPath) !== -1) && (!file.endsWith(path.sep)))
+  );
+
+  const instructions = filtered.map(file => {
+    return {
+      type: 'copy',
+      source: file,
+      destination: path.join(RELOADED_PATH, file.substr(idx)),
+    };
+  });
+  instructions.push(setModTypeInstruction);
+  return Promise.resolve({ instructions });
+}
+
 // MAIN FUNCTIONS ///////////////////////////////////////////////////////////////
 
-//Notify User of Setup instructions for Mod Managers
+//Notify User of Setup instructions for Reloaded-II
 function setupNotify(api) {
-  const NOTIF_ID = `${GAME_ID}-setup-notification`;
+  const NOTIF_ID = `${GAME_ID}-setup`;
   const MESSAGE = 'Reloaded Mod Manager Setup Required';
   api.sendNotification({
     id: NOTIF_ID,
@@ -477,7 +518,7 @@ function setupNotify(api) {
         title: 'More',
         action: (dismiss) => {
           api.showDialog('question', MESSAGE, {
-            text: 'The Reloaded Mod Manager tool downloaded by this extension requires setup.\n'
+            text: 'The Reloaded-II Mod Manager tool downloaded by this extension requires setup.\n'
                 + '\n'
                 + `Please launch the tool and set the location of the ${GAME_NAME} executable.\n`
                 + '\n'
@@ -485,7 +526,20 @@ function setupNotify(api) {
                 + '\n'
                 + 'You must launch the game from Reloaded for mods to load in the game.\n'
                 + '\n'
+                + '\n'
+                + `IMPORTANT: You may need to run Reloaded-II as Administrator for setup and dependency updates.\n`
+                + '\n'
+                + 'If you get a message about missing mod dependencies or .NET Runtimes, you may need to run Reloaded as an Administrator to allow dependencies to download.\n'
+                + '\n'
+                + 'You can launch Reloaded as Admin using the button below.\n'
+                + '\n'
           }, [
+            {
+              label: `Run Reloaded-II as Admin`, action: () => {
+                runReloadedAdmin(api);
+                dismiss();
+              }
+            },
             { label: 'Acknowledge', action: () => dismiss() },
             {
               label: 'Never Show Again', action: () => {
@@ -500,20 +554,43 @@ function setupNotify(api) {
   });    
 }
 
+//Run Reloaded-II as Admin (allow dependency download in priviledged folders)
+function runReloadedAdmin(api) {
+  const TOOL_ID = "ReloadedModManager";
+  const TOOL_NAME = RELOADED_NAME;
+  const state = api.store.getState();
+  const tool = util.getSafe(state, ['settings', 'gameMode', 'discovered', GAME_ID, 'tools', TOOL_ID], undefined);
+
+  try {
+    const TOOL_PATH = tool.path;
+    if (TOOL_PATH !== undefined) {
+      return api.runExecutable(path.join(util.getVortexPath('application'), 'resources', 'elevate.exe'), [TOOL_PATH], { suggestDeploy: false, detached: true })
+        .catch(err => api.showErrorNotification(`Failed to run ${TOOL_NAME} as Admin`, err,
+          { allowReport: ['EPERM', 'EACCESS', 'ENOENT'].indexOf(err.code) !== -1 })
+        );
+    }
+    else {
+      return api.showErrorNotification(`Failed to run ${TOOL_NAME} as Admin`, `Path to ${TOOL_NAME} executable could not be found. Ensure ${TOOL_NAME} is installed through Vortex.`);
+    }
+  } catch (err) {
+    return api.showErrorNotification(`Failed to run ${TOOL_NAME} as Admin`, err, { allowReport: ['EPERM', 'EACCESS', 'ENOENT'].indexOf(err.code) !== -1 });
+  }
+}
+
 //Setup function
 async function setup(discovery, api, gameSpec) {
   // SYNCHRONOUS CODE ////////////////////////////////////
+  GAME_PATH = discovery.path;
   setupNotify(api);
   const state = api.getState();
-  GAME_PATH = discovery.path;
   STAGING_FOLDER = selectors.installPathForGame(state, GAME_ID);
   DOWNLOAD_FOLDER = selectors.downloadPathForGame(state, GAME_ID);
   // ASYNC CODE //////////////////////////////////////////
-  await fs.ensureDirWritableAsync(path.join(discovery.path, RELOADEDMODLOADER_PATH));
-  await fs.ensureDirWritableAsync(path.join(discovery.path, SAVE_PATH));
+  await fs.ensureDirWritableAsync(path.join(GAME_PATH, RELOADEDMODLOADER_PATH));
+  await fs.ensureDirWritableAsync(path.join(GAME_PATH, SAVE_PATH));
   await downloadModManager(api, gameSpec);
   return fs.ensureFileAsync(
-    path.join(discovery.path, RELOADED_PATH, "portable.txt")
+    path.join(GAME_PATH, RELOADED_PATH, "portable.txt")
   );
 }
 
@@ -527,25 +604,10 @@ function applyGame(context, gameSpec) {
     requiresLauncher: requiresLauncher,
     setup: async (discovery) => await setup(discovery, context.api, gameSpec),
     executable: () => gameSpec.game.executable,
-    //getGameVersion: resolveGameVersion,
     supportedTools: [
-      /*
-      {
-        id: "LaunchModdedGame",
-        name: "Launch Modded Game",
-        logo: "exec.png",
-        executable: () => RELOADED_EXEC,
-        requiredFiles: [RELOADED_EXEC],
-        detach: true,
-        relative: true,
-        exclusive: true,
-        defaultPrimary: true,
-        parameters: [`--launch "${path.join(GAME_PATH, EXEC)}"`]
-      },
-      */
       {
         id: "ReloadedModManager",
-        name: "Reloaded Mod Manager",
+        name: RELOADED_NAME,
         logo: "reloaded.png",
         executable: () => RELOADED_EXEC,
         requiredFiles: [RELOADED_EXEC],
@@ -612,9 +674,62 @@ function applyGame(context, gameSpec) {
 function main(context) {
   applyGame(context, spec);
   context.once(() => { // put code here that should be run (once) when Vortex starts up
-
+    context.api.onAsync('did-deploy', async (profileId, deployment) => {
+      const lastActiveProfile = selectors.lastActiveProfileForGame(context.api.getState(), GAME_ID);
+      if (profileId !== lastActiveProfile) return;
+      return deployNotify(context.api);
+    });
   });
   return true;
+}
+
+//Notify User to run Mod Merger Utility after deployment
+function deployNotify(api) {
+  const NOTIF_ID = `${GAME_ID}-deploy`;
+  const MOD_NAME = RELOADED_NAME;
+  const MESSAGE = `Run ${MOD_NAME} as Admin`;
+  api.sendNotification({
+    id: NOTIF_ID,
+    type: 'warning',
+    message: MESSAGE,
+    allowSuppress: true,
+    actions: [
+      {
+        title: 'Run Reloaded (Admin)',
+        action: (dismiss) => {
+          runReloadedAdmin(api);
+          dismiss();
+        },
+      },
+      {
+        title: 'More',
+        action: (dismiss) => {
+          api.showDialog('question', MESSAGE, {
+            text: `If your game is installed in a protected folder, such as "C:\Program Files (x86)" (default for Steam), you must run ${MOD_NAME} as Administrator.\n`
+                + '\n'
+                + `Use the button below to launch ${MOD_NAME} as Administrator.\n`
+                + '\n'
+                + `If your game is NOT installed in a priviledged folder, you can suppress this notification with the "Never Show Again" button below.\n`
+                + '\n'
+          }, [
+            {
+              label: 'Run Reloaded (Admin)', action: () => {
+                runReloadedAdmin(api);
+                dismiss();
+              }
+            },
+            { label: 'Continue', action: () => dismiss() },
+            {
+              label: 'Never Show Again', action: () => {
+                api.suppressNotification(NOTIF_ID);
+                dismiss();
+              }
+            },
+          ]);
+        },
+      },
+    ],
+  });
 }
 
 //export to Vortex