+++ game-farcryprimal/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - None Planned
 
-## [0.1.0] - 2025-XX-XX
+## [0.1.1] - 2025-10-14
+
+- Fixed button to open Saves folder.
+
+## [0.1.0] - 2025-10-04
 
 - Initial release